# docs/swagger.yaml
openapi: 3.0.2
info:
  title: Games-Api
  version: 1.0.0
  description: This is Swagger UI for Games API
paths:
  /games:
    get:
      tags:
        - games
      summary: Get a list of all games
      description: Return a list of all games in API
      responses:
        '200':
          description: List of games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/game'
              examples:
                listOfGames:
                  value:
                    - id: 1
                      name: "Witcher 3"
                      price: 29.99
                    - id: 2
                      name: "Cyberpunk 2077"
                      price: 59.99
        '400':
          description: Invalid request
    post:
      tags:
        - games
      summary: Add a new game
      description: Add a new game to the list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/new-game'
      responses:
        '201':
          description: Game created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/game'
  /games/{id}:
    get:
      tags:
        - games
      summary: Get a game by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Game details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/game'
        '404':
          description: Game not found
    delete:
      tags:
        - games
      summary: Delete a game by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Game deleted
        '404':
          description: Game not found
components:
  schemas:
    game:
      type: object
      required:
        - id
        - name
        - price
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
          format: float
      example:
        id: 1
        name: "Witcher 3"
        price: 29.99
    new-game:
      type: object
      required:
        - name
        - price
      properties:
        name:
          type: string
        price:
          type: number
          format: float
      example:
        name: "Tetris"
        price: 0.99